rename SoundPeriod to TimePeriod
