+++ src/Domain/Notifications/Channel.php
@@ -3,6 +3,7 @@
 namespace ArtARTs36\MergeRequestLinter\Domain\Notifications;
 
 use ArtARTs36\MergeRequestLinter\Shared\Contracts\DataStructure\Map;
+use ArtARTs36\MergeRequestLinter\Shared\Time\TimePeriod;
 
 /**
  * @codeCoverageIgnore
@@ -14,8 +15,8 @@ class Channel
      */
     public function __construct(
         public readonly ChannelType $type,
-        public readonly Map $params,
-        public readonly SoundPeriod $sound,
+        public readonly Map         $params,
+        public readonly TimePeriod  $sound,
     ) {
         //
     }
