+++ src/Infrastructure/Configuration/Loader/Mapper/NotificationsMapper.php
@@ -6,11 +6,11 @@ use ArtARTs36\MergeRequestLinter\Domain\Configuration\NotificationEventMessage;
 use ArtARTs36\MergeRequestLinter\Domain\Configuration\NotificationsConfig;
 use ArtARTs36\MergeRequestLinter\Domain\Notifications\Channel;
 use ArtARTs36\MergeRequestLinter\Domain\Notifications\ChannelType;
-use ArtARTs36\MergeRequestLinter\Domain\Notifications\SoundPeriod;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Configuration\Exceptions\ConfigInvalidException;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Configuration\ConfigValueTransformer;
 use ArtARTs36\MergeRequestLinter\Shared\Contracts\DataStructure\Map;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\ArrayMap;
+use ArtARTs36\MergeRequestLinter\Shared\Time\TimePeriod;
 
 class NotificationsMapper
 {
@@ -130,7 +130,7 @@ class NotificationsMapper
 
         foreach ($data as $name => $channel) {
             $type = ChannelType::from($channel['type']);
-            $sound = SoundPeriod::make($channel['sound_at'] ?? '');
+            $sound = TimePeriod::make($channel['sound_at'] ?? '');
 
             unset($channel['type']);
             unset($channel['sound_at']);
